Refactored common elements into TaskCommonDetail
