+++ frontend/public/locales/en/translation.json
@@ -55,20 +55,18 @@
         "button_load_more": "Load more",
         "button_add_task": "Add new task",
         "button_close_add_task": "Close",
+        "button_add": "Add",
+        "button_save": "Save",
         "loading": "Loading...",
+        "none": "none",
         "error_load_project_and_drawer": "Couldn't load project and drawers 😢",
         "error_load_task": "Couldn't load tasks 😢",
-        "create": {
-            "name_placeholder": "Enter a name",
-            "button_add": "Add",
-            "none": "none",
-            "task_create_success": "Created a new task.",
-            "task_create_error": "Failed to create a new task.",
-            "task_create_no_name": "Write a name of a task."
-        },
+        "name_placeholder": "Enter a name",
+        "task_create_success": "Created a new task.",
+        "task_create_error": "Failed to create a new task.",
+        "task_create_no_name": "Write a name of a task.",
         "edit": {
-            "display": "Edit",
-            "name_change_success": "The name has changed."
+            "display": "Edit"
         },
         "delete": {
             "display": "Delete",
