+++ frontend/src/components/project/taskDetails/TaskCreateElement.jsx
@@ -4,13 +4,12 @@ import ModalWindow from "@components/common/ModalWindow"
 import TaskCreate from "@components/project/taskDetails/TaskCreate"
 
 const TaskCreateElement = () => {
-    /* 현재 url 기준으로 useState default값 정하기*/
     const navigate = useNavigate()
 
-    const [projectId] = useOutletContext()
+    const [projectID] = useOutletContext()
 
     const closeCreate = () => {
-        navigate(`/app/projects/${projectId}`)
+        navigate(`/app/projects/${projectID}`)
     }
 
     return (
