+++ frontend/src/components/project/taskDetails/TaskDetailElement.jsx
@@ -4,7 +4,6 @@ import ModalWindow from "@components/common/ModalWindow"
 import TaskDetail from "@components/project/taskDetails/TaskDetail"
 
 const TaskDetailElement = () => {
-    /* 현재 url 기준으로 useState default값 정하기*/
     const navigate = useNavigate()
 
     const [projectID] = useOutletContext()
